+++ components/scoreboard/TeamRow.js
@@ -1,14 +1,19 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import styled from 'styled-components';
+import { darken } from 'polished';
+import { observer } from 'mobx-react';
 import Animated from 'animated/lib/targets/react-dom';
 
 const Row = styled.div`
   width: 100%;
-  height: 38px;
+  height: 50px;
   display: flex;
   align-items: center;
-  justify-content: space-between;
+  justify-content: flex-start;
   box-sizing: content-box;
+  flex-direction: row;
+  overflow: hidden;
 `;
 
 const ImageLogo = styled.img`
@@ -17,73 +22,61 @@ const ImageLogo = styled.img`
 `;
 
 const ShirtColor = styled.div`
-  border-radius: 3px;
+  border-radius: 0px;
   background-color: ${props => props.hex};
-  width: 16px;
-  height: 16px;
-  margin-left: 8px;
-  border: 1px solid #111;
+  background: linear-gradient(
+    ${props => darken(0.15, props.hex)},
+    ${props => props.hex},
+    ${props => darken(0.15, props.hex)}
+  );
+  width: 8px;
+  height: 50px;
 `;
 
 const Name = styled.div`
   letter-spacing: 1px;
   font-size: 20px;
   min-width: 100px;
-  max-width: 170px;
+  max-width: 220px;
   text-transform: uppercase;
   margin-left: 8px;
-  text-shadow: 1px 1px #000;
   height: 22px;
   line-height: 22px;
   text-overflow: ellipsis;
   overflow: hidden;
   white-space: nowrap;
-`;
-
-const Sets = styled.div`
-  width: 18px;
-  padding: 2px 0px;
-  text-align: center;
-  color: #000;
-  background-color: papayawhip;
-  border-radius: 3px;
-  height: 20px;
-`;
-
-const Score = styled.div`
-  font-weight: 300;
-  background-color: white;
-  color: black;
-  width: 34px;
-  height: 34px;
-  line-height: 34px;
-  font-size: 24px;
-  text-align: center;
-  border-radius: 3px;
-  margin-left: 8px;
-`;
-
-const LeftContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-`;
-
-const RightContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
+  color: ${props => props.color};
+  text-shadow: 1px 1px #000;
+  font-weight: 400;
 `;
 
 const AnimatedLogo = Animated.createAnimatedComponent(ImageLogo);
 const AnimatedShirtColor = Animated.createAnimatedComponent(ShirtColor);
 
 class TeamRow extends React.Component {
+  static propTypes = {
+    color: PropTypes.string,
+    textColor: PropTypes.string,
+    logo: PropTypes.string,
+    name: PropTypes.string,
+    showColor: PropTypes.bool,
+    showLogo: PropTypes.bool,
+  };
+
+  static defaultProps = {
+    logo: '',
+    name: '',
+    color: '',
+    showLogo: false,
+    showColor: false,
+    textColor: '#ffffff',
+  };
+
   constructor(props) {
     super(props);
     this.state = {
-      logoAnim: new Animated.Value(0),
-      colorAnim: new Animated.Value(0),
+      logoAnim: new Animated.Value(props.showLogo ? 1 : 0),
+      colorAnim: new Animated.Value(props.showColor ? 1 : 0),
     };
   }
 
@@ -91,10 +84,8 @@ class TeamRow extends React.Component {
     return (
       this.props.color !== nextProps.color ||
       this.props.name !== nextProps.name ||
-      this.props.points !== nextProps.points ||
-      this.props.sets !== nextProps.sets ||
       this.props.showColor !== nextProps.showColor ||
-      this.props.showLogos !== nextProps.showLogos
+      this.props.showLogo !== nextProps.showLogo
     );
   }
 
@@ -102,7 +93,7 @@ class TeamRow extends React.Component {
     if (!nextProps.showColor) {
       Animated.spring(this.state.colorAnim, { toValue: 0 }).start();
     }
-    if (!nextProps.showLogos) {
+    if (!nextProps.showLogo) {
       Animated.spring(this.state.logoAnim, { toValue: 0 }).start();
     }
   }
@@ -111,7 +102,7 @@ class TeamRow extends React.Component {
     if (this.props.showColor) {
       Animated.spring(this.state.colorAnim, { toValue: 1 }).start();
     }
-    if (this.props.showLogos) {
+    if (this.props.showLogo) {
       Animated.spring(this.state.logoAnim, { toValue: 1 }).start();
     }
   }
@@ -119,61 +110,36 @@ class TeamRow extends React.Component {
   render() {
     return (
       <Row>
-        <LeftContainer>
-          <AnimatedLogo
-            style={{
-              opacity: this.state.logoAnim,
-              width: this.state.logoAnim.interpolate({
-                inputRange: [0, 1],
-                outputRange: [0, 33],
-              }),
-            }}
-            src={this.props.logo}
-            alt="Home Team Logo"
-          />
-          <AnimatedShirtColor
-            style={{
-              opacity: this.state.colorAnim,
-              width: this.state.colorAnim.interpolate({
-                inputRange: [0, 1],
-                outputRange: [0, 16],
-              }),
-              marginLeft: this.state.colorAnim.interpolate({
-                inputRange: [0, 1],
-                outputRange: [0, 8],
-              }),
-            }}
-            hex={this.props.color}
-          />
-          <Name>{this.props.name}</Name>
-        </LeftContainer>
-        <RightContainer>
-          <Sets>{this.props.sets}</Sets>
-          <Score>{this.props.points}</Score>
-        </RightContainer>
+        <AnimatedShirtColor
+          style={{
+            opacity: this.state.colorAnim,
+            width: this.state.colorAnim.interpolate({
+              inputRange: [0, 1],
+              outputRange: [0, 8],
+            }),
+          }}
+          hex={this.props.color}
+        />
+        <AnimatedLogo
+          style={{
+            opacity: this.state.logoAnim,
+            width: this.state.logoAnim.interpolate({
+              inputRange: [0, 1],
+              outputRange: [0, 33],
+            }),
+            marginLeft: this.state.logoAnim.interpolate({
+              inputRange: [0, 1],
+              outputRange: [0, 8],
+            }),
+          }}
+          src={this.props.logo}
+          alt="Team Logo"
+        />
+
+        <Name color={this.props.textColor}>{this.props.name}</Name>
       </Row>
     );
   }
 }
 
-TeamRow.propTypes = {
-  color: PropTypes.string,
-  logo: PropTypes.string,
-  name: PropTypes.string,
-  points: PropTypes.number,
-  sets: PropTypes.number,
-  showColor: PropTypes.bool,
-  showLogos: PropTypes.bool,
-};
-
-TeamRow.defaultProps = {
-  sets: 0,
-  points: 0,
-  logo: '',
-  name: '',
-  color: '',
-  showLogos: false,
-  showColor: false,
-};
-
-export default TeamRow;
+export default observer(TeamRow);
