+++ components/scoreboard/color.js
