+++ package.json
@@ -10,7 +10,8 @@
     "install-redis": "./lib/redis/install.sh",
     "redis": "./redis/src/redis-server",
     "lint": "node ./node_modules/eslint/bin/eslint.js ./components/** -c ./.eslintrc.json --fix",
-    "prettier": "prettier-eslint --single-quote --trailing-comma all '{components,pages,src}/**/*.js' --write"
+    "prettier":
+      "prettier-eslint --single-quote --trailing-comma all '{components,pages,src}/**/*.js' --write"
   },
   "engines": {
     "node": "8.2.1",
@@ -45,11 +46,11 @@
     "@storybook/react": "3.2.8",
     "babel-eslint": "^7.2.2",
     "babel-preset-env": "^1.2.2",
-    "eslint": "3.19.0",
-    "eslint-config-airbnb": "^14.1.0",
-    "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "4.0.0",
-    "eslint-plugin-react": "6.9.0",
+    "eslint": "^4.6.0",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-jsx-a11y": "5.1.1",
+    "eslint-plugin-react": "7.2.0",
     "prettier": "^1.5.2",
     "prettier-eslint": "^6.4.2",
     "prettier-eslint-cli": "^4.1.1"
