+++ pages/firebase.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Head from 'next/head';
 import { Provider } from 'mobx-react';
-import Score from '../components/scoreboard/new/Scoreboard';
+import Score from '../components/scoreboard/Scoreboard';
 import FirebaseStore from '../store/firebase';
 
 class Scoreboard extends React.Component {
